expressvis-server/requirements-old.txt
uwsgi
Django==3.2.5
django-cors-headers
djangorestframework==3.12.4
fastcluster
feather==0.1.2
numpy==1.20.2
pandas==1.3.3
PyJWT==2.3.0
pymongo==3.12.0
requests==2.25.1
scikit_learn==1.0.1
scipy==1.7.1
statsmodels==0.13.1
